[PATCH] move entry point to src/index.js
