+++ src/Passable.js
@@ -4,7 +4,6 @@ import enforce from './enforce';
 import passRunner from './pass_runner';
 import { passableArgs, initResponseObject, initField, onFail, root, runtimeError } from 'Helpers';
 import { Errors } from 'Constants';
-import { version } from '../version.json';
 
 const FAIL: Severity = 'fail';
 
@@ -69,7 +68,4 @@ class Passable {
     }
 }
 
-const passable: Function = (name: string, ...args: PassableArguments) => new Passable(name, ...args);
-passable.VERSION = version;
-
-export default passable;
+export default Passable;
