+++ webpack.config.js
@@ -24,7 +24,7 @@ if (env === 'build') {
 }
 
 const config = {
-    entry: `${__dirname}/src/Passable.js`,
+    entry: `${__dirname}/src/index.js`,
     devtool: 'source-map',
     output: {
         path: `${__dirname}/${outputDir}`,
